api/requirements.txt
fastapi
torch
pandas
pillow
scikit-learn
transformers
uvicorn
timm
einops
tabulate
deep_translator
langdetect
numpy